# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for LLM trace logging, testing, and regression detection",
        "title": "Regrada API",
        "contact": {
            "name": "API Support",
            "email": "support@regrada.ai"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/auth/signup": {
            "post": {
                "description": "Register a new user account with optional organization creation or invite token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Signup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "create_organization": {
                                    "type": "boolean"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "invite_token": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "organization_name": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "organization_id": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "user_confirmed": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}